backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
asgiref==3.8.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
django-jazzmin==3.0.0
django-environ==0.11.2
django-apscheduler==0.6.2
six==1.16.0